requirements.txt
PyQt5>=5.15.0
mido>=1.2.0
python-rtmidi>=1.4.0
